+++ Resources/settingsController.js
@@ -1,14 +1,13 @@
 // settings CONTROLLER
+Ti.include('setThemeView.js');
+Ti.include('setVolumeView.js');
+Ti.include('aboutView.js');
+
 settingsList.addEventListener('click', function(e) {
-	if (sound) {
+	if (mute === false) {
 		swosh.play();
 	}
 	if (e.index === 0) {
-		Ti.include('setThemeView.js');
-		//set background gradient
-		themeWin.backgroundGradient = {
-			colors: ["#C7D6E9", "#fff"]
-		};
 		// open the tab with a default slide animation
 		tabGroup.activeTab.open(themeWin, {
 			animation: true,
@@ -16,22 +15,12 @@ settingsList.addEventListener('click', function(e) {
 
 	}
 	if (e.index === 1) {
-		Ti.include('setVolumeView.js');
-		//set background gradient
-		volumeWin.backgroundGradient = {
-			colors: ["#C7D6E9", "#fff"]
-		};
 		// open the tab with a default slide animation
 		tabGroup.activeTab.open(volumeWin, {
 			animation: true,
 		});
 	}
 	if (e.index === 2) {
-		Ti.include('aboutView.js');
-		//set background gradient
-		aboutWin.backgroundGradient = {
-			colors: ["#C7D6E9", "#fff"]
-		};
 		// open the tab with a default slide animation
 		tabGroup.activeTab.open(aboutWin, {
 			animation: true,